modify meal plan menu

+++ lib/features/meal_plan/data/repositories/meal_plan_repository_impl.dart
@@ -333,7 +333,7 @@ class MealPlanRepositoryImpl implements MealPlanRepository {
     required AddMealCategoryOption mealCategory,
     required AddMealAiRecipe recipe,
     required String source,
-    required int servingCount,
+    required double servingCount,
   }) async {
     try {
       // Delegate to remote data source.
